shared: move dirents cleanup from private-tree.h to shared/cleanup.h
struct dirents, cleanup_dirents(), and __cleanup_dirents are currently
defined in libnvme/src/nvme/private-tree.h, limiting their use to
libnvme internals.

Move them to shared/cleanup.h (adding the required #include <dirent.h>)
so that nvme-cli plugins can use the same cleanup pattern without
duplicating the logic. private-tree.h already includes cleanup.h via
its existing #include "cleanup.h", so no functional change there.
